Rolling Dice using Arduino and 7 Segment Display
How To Make Rolling Dice using Arduino and 7 Segment Display
---------------------------------------------------------------------------------------------------------------------
In this video,How to make Rolling Dice using Arduino and 7 Segment Display.I hope you will like this video & Subscribe The Channel.

Topics Covered:-
1] How To Make Rolling Dice using Arduino and 7 Segment Display.
2] Which Components Required.
3] Design the circuit.
4] Codeing Part.
5] Testing The Circuit.

we will implement a rolling dice function where we will use a button to display a random number between 1 and 6.

In the practical implementation, I connected the current limiting resistors to the individual LEDs of the 7 Segment Display. 

Working:-
The working of the project is similar to the above project except that a Push Button is interfaced to Arduino additionally. When the button is pushed, you can see Arduino displaying numbers from 0 to 9. The moment you release the button, a random number between 1and 6 is displayed on the 7 Segment Display.

Components Required

1] Arduino UNO
2] 7 Segment Display
3] Push Button
4] 7 x 220Ω Resistors (1/4 Watt)
5] Breadboard
6] Power Supply
7] Connecting Wires

# Rolling Dice.
# 7 Segment Display.
# Arduino Nano.
#Button.